+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- `State` -> `Model`
+- `Init` -> `ModelInit`
+- `StateUpdate` -> `ModelUpdate`
+- `Subscribe` -> `ModelSubscribe`
+- `SubscriptionMap` -> `Subscriptions`
+- `StateInstance` -> `State`
+- `Storage`.`stateDefinition` -> `getModel`
+- React `StateData`.`state` -> `model`
 
 ## [0.2.0] - 2019-06-19
 ### Added